requirements.txt
click==8.1.3
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
SQLAlchemy==1.4.39
Werkzeug==2.1.2
gunicorn==20.1.0
pytz==2022.1
psycopg2==2.9.3
python-dotenv==0.20.0